Adding default_post_time_range_seconds columns to local_user and site.

+++ crates/api_common/src/comment.rs
@@ -119,7 +119,7 @@ pub struct GetComments {
   #[cfg_attr(feature = "full", ts(optional))]
   /// Filter to within a given time range, in seconds.
   /// IE 60 would give results for the past minute.
-  pub time_range_seconds: Option<i64>,
+  pub time_range_seconds: Option<i32>,
   #[cfg_attr(feature = "full", ts(optional))]
   pub max_depth: Option<i32>,
   #[cfg_attr(feature = "full", ts(optional))]
