+++ crates/api_common/src/post.rs
@@ -95,7 +95,7 @@ pub struct GetPosts {
   #[cfg_attr(feature = "full", ts(optional))]
   /// Filter to within a given time range, in seconds.
   /// IE 60 would give results for the past minute.
-  pub time_range_seconds: Option<i64>,
+  pub time_range_seconds: Option<i32>,
   /// DEPRECATED, use page_cursor
   #[cfg_attr(feature = "full", ts(optional))]
   pub page: Option<i64>,
